+++ ts/src/common.ts
@@ -1,7 +1,6 @@
 import {
   AgentPubKey,
   AppInfo,
-  authorizeSigningCredentials,
   CallZomeResponse,
   Cell,
   CellType,
@@ -93,8 +92,6 @@ export const enableAndGetAgentApp = async (
     agentPubKey,
     cells,
     namedCells,
-    authorizeSigningCredentials: (cellId, functions) =>
-      authorizeSigningCredentials(conductor.adminWs(), cellId, functions),
   };
   return agentApp;
 };